Create a new class for the image view widget
